backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-Mail==0.9.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.74
openai==0.27.8
